refactor: 초기 데이터 fetch로직 분리
