+++ CurrencyConverterApp/CurrencyConverterApp/Domain/Entity/ExchangeRateInfo.swift
