+++ CurrencyConverterApp/CurrencyConverterApp/Network/ExchangeRateResponse.swift
@@ -5,9 +5,6 @@
 //  Created by 박주성 on 4/17/25.
 //
 
-
-// ExchangeRateResponse.swift
-
 struct ExchangeRateResponse {
     let exchangeRateList: [ExchangeRateInfo]
     let timeLastUpdateUnix: Int
